third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEDisplacementMapElement-svgdom-scale-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'scale' property of the SVGFEDisplacementMapElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS displacementMap.scale.baseVal is 34
PASS displacementMap.scale.baseVal is 64
PASS successfullyParsed is true

TEST COMPLETE
